+++ src/Loader.php
@@ -55,6 +55,7 @@ class Loader
         $this->register_service(\AperturePro\Services\Admin::class);
         $this->register_service(\AperturePro\Services\ClientPortal::class);
         $this->register_service(\AperturePro\Services\REST::class);
+        $this->register_service(\AperturePro\Services\Email::class);
     }
 
     /**